+++ interceptor/proxy_handlers.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/kedacore/http-add-on/interceptor/config"
 	"github.com/kedacore/http-add-on/interceptor/handler"
+	"github.com/kedacore/http-add-on/pkg/k8s"
 	kedanet "github.com/kedacore/http-add-on/pkg/net"
 	"github.com/kedacore/http-add-on/pkg/util"
 )
@@ -52,6 +53,8 @@ func newForwardingHandler(
 	fwdCfg forwardingConfig,
 	tlsCfg *tls.Config,
 	tracingCfg *config.Tracing,
+	placeholderHandler *handler.PlaceholderHandler,
+	endpointsCache k8s.EndpointsCache,
 ) http.Handler {
 	roundTripper := &http.Transport{
 		Proxy:                 http.ProxyFromEnvironment,
@@ -68,6 +71,25 @@ func newForwardingHandler(
 		var uh *handler.Upstream
 		ctx := r.Context()
 		httpso := util.HTTPSOFromContext(ctx)
+
+		if placeholderHandler != nil && httpso.Spec.PlaceholderConfig != nil && httpso.Spec.PlaceholderConfig.Enabled {
+			endpoints, err := endpointsCache.Get(httpso.GetNamespace(), httpso.Spec.ScaleTargetRef.Service)
+			if err == nil && workloadActiveEndpoints(endpoints) == 0 {
+				lggr.Info("serving placeholder page immediately",
+					"namespace", httpso.GetNamespace(),
+					"service", httpso.Spec.ScaleTargetRef.Service,
+					"reason", "zero replicas detected")
+
+				if placeholderErr := placeholderHandler.ServePlaceholder(w, r, httpso); placeholderErr != nil {
+					lggr.Error(placeholderErr, "failed to serve placeholder page")
+					w.WriteHeader(http.StatusBadGateway)
+					if _, err := w.Write([]byte("error serving placeholder page")); err != nil {
+						lggr.Error(err, "could not write error response to client")
+					}
+				}
+				return
+			}
+		}
 
 		waitFuncCtx, done := context.WithTimeout(ctx, fwdCfg.waitTimeout)
 		defer done()